Moved hooks and renamed stores to utils
