+++ frontend/src/features/data_entry/components/check_and_save/CheckAndSaveForm.test.tsx
@@ -11,10 +11,10 @@ import {
   PollingStationDataEntrySaveHandler,
 } from "@/testing/api-mocks";
 
-import { DataEntryProvider } from "../../stores/DataEntryProvider";
 import { errorWarningMocks, getDefaultFormSection, getEmptyDataEntryRequest } from "../../testing/mock-data";
 import { overrideServerClaimDataEntryResponse } from "../../testing/test.utils";
 import { DataEntryState } from "../../types/types";
+import { DataEntryProvider } from "../DataEntryProvider";
 import { CheckAndSaveForm } from "./CheckAndSaveForm";
 
 function getDefaultValues() {
