+++ frontend/src/features/data_entry/components/differences/DifferenceForm.test.tsx
@@ -9,7 +9,6 @@ import {
   PollingStationDataEntrySaveHandler,
 } from "@/testing/api-mocks";
 
-import { DataEntryProvider } from "../../stores/DataEntryProvider";
 import { errorWarningMocks, getDefaultFormSection, getEmptyDataEntryRequest } from "../../testing/mock-data";
 import {
   expectFieldsToBeInvalidAndToHaveAccessibleErrorMessage,
@@ -19,6 +18,7 @@ import {
   overrideServerClaimDataEntryResponse,
 } from "../../testing/test.utils";
 import { DataEntryState } from "../../types/types";
+import { DataEntryProvider } from "../DataEntryProvider";
 import { DifferencesForm } from "./DifferencesForm";
 
 const defaultDataEntryState: DataEntryState = {
