+++ frontend/src/features/data_entry/components/voters_and_votes/VotersAndVotesForm.test.tsx
@@ -9,8 +9,6 @@ import {
   PollingStationDataEntrySaveHandler,
 } from "@/testing/api-mocks";
 
-import { DataEntryProvider } from "../../stores/DataEntryProvider";
-import { getClientState } from "../../stores/dataEntryUtils";
 import { errorWarningMocks, getDefaultFormSection, getEmptyDataEntryRequest } from "../../testing/mock-data";
 import {
   expectFieldsToBeInvalidAndToHaveAccessibleErrorMessage,
@@ -20,6 +18,8 @@ import {
   overrideServerClaimDataEntryResponse,
 } from "../../testing/test.utils";
 import { DataEntryState } from "../../types/types";
+import { getClientState } from "../../utils/dataEntryUtils";
+import { DataEntryProvider } from "../DataEntryProvider";
 import { VotersAndVotesForm } from "./VotersAndVotesForm";
 
 const initialValues: PollingStationResults = {
