+++ frontend/src/features/data_entry/components/voters_and_votes/useVotersAndVotes.ts
@@ -1,4 +1,4 @@
-import { useDataEntryFormSection } from "../../stores/useDataEntryFormSection";
+import { useDataEntryFormSection } from "../../hooks/useDataEntryFormSection";
 import { SubmitCurrentFormOptions } from "../../types/types";
 import {
   formValuesToValues,
