+++ frontend/src/features/data_entry/hooks/useInitialDataEntryState.ts
