+++ frontend/src/features/data_entry/utils/ValidationResults.test.ts
