+++ frontend/src/features/data_entry/utils/ValidationResults.ts
