+++ frontend/src/features/data_entry/utils/dataEntryUtils.test.ts
