+++ frontend/src/features/data_entry/utils/dataEntryUtils.ts
