+++ frontend/src/features/data_entry/utils/reducer.test.ts
