+++ frontend/src/features/data_entry/utils/reducer.ts
