+++ frontend/src/features/data_entry/utils/utils.ts
